+++ src/client/components/LinksTable/Link/Link.tsx
@@ -7,9 +7,9 @@ import { CountryIso } from 'meta/area/countryIso'
 import { Link as LinkType } from 'meta/cycleData/links/link'
 import { Objects } from 'utils/objects'
 
-import { LinksActions } from 'client/store/admin/links/actions'
-import { useIsVerificationInProgress } from 'client/store/admin/links/hooks/verification'
 import { useAppDispatch } from 'client/store/hooks'
+import { LinksActions } from 'client/store/links/actions'
+import { useIsVerificationInProgress } from 'client/store/links/hooks/verification'
 import { useSectionRouteParams } from 'client/hooks/routeParams'
 import Button, { ButtonSize, ButtonType } from 'client/components/Buttons/Button'
 import LinkCommon from 'client/components/Links/Link'
