--- src/client/store/admin/name.ts
@@ -1,1 +1,0 @@
-export const AdminSliceName = 'admin'
