--- src/client/store/admin/slice/index.ts
@@ -1,5 +1,0 @@
-import { combineSlices } from '@reduxjs/toolkit'
-
-import { LinksSlice } from 'client/store/admin/links/slice'
-
-export const AdminSlice = combineSlices(LinksSlice)
