+++ src/client/store/links/actions/reset.ts
@@ -1,0 +1,3 @@
+import { createAction } from '@reduxjs/toolkit'
+
+export const reset = createAction('links/reset')
